+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1-beta
+##### 2017.11.28
+* New payload parameter
+* Passing decoded token in parameter
+
 ## 1.0.0-beta
 ##### 2017.11.20
 * Major bug fix